+++ src/main/java/com/sparta/engineering72/sakilaproject/entities/Staff.java
@@ -14,7 +14,7 @@ public class Staff implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="staff_id", unique=true, nullable=false)
-	private byte staffId;
+	private int staffId;
 
 	@Column(nullable=false)
 	private byte active;
@@ -65,11 +65,11 @@ public class Staff implements Serializable {
 	public Staff() {
 	}
 
-	public byte getStaffId() {
+	public int getStaffId() {
 		return this.staffId;
 	}
 
-	public void setStaffId(byte staffId) {
+	public void setStaffId(int staffId) {
 		this.staffId = staffId;
 	}
